update.c: fix LP64 int x,y,rightshim → long in updateValue, numbers now visible in cells
FossilOrigin-Name: b45f4a9d89e3adb1accc0ef094c31add852bfcaed574dc6de06e5bffc404edb0

+++ src/atk/table/update.c
@@ -432,10 +432,10 @@ int precision;
 struct rectangle *cellBounds;
 {
     struct table *T = MyTable(V);
-    int     x, y;
+    long    x, y;
     double val;
     int decpt, sign;
-    int rightshim;
+    long rightshim;
     char    buf[1030], *p;
     int adjustedprecision = precision + 1;
